+++ Source/Lib/Codec/EbSequenceControlSet.c
@@ -366,7 +366,7 @@ extern EbErrorType eb_vp9_sb_params_init(
                 EB_FALSE :
                 EB_TRUE;
 
-           sequence_control_set_ptr->sb_params_array[sb_index].ep_scan_block_valid_block[md_scan_block_index] = (uint32_t)~0;
+           sequence_control_set_ptr->sb_params_array[sb_index].ep_scan_block_valid_block[md_scan_block_index] = (uint16_t)~0;
         }
 
         // Find the valid block for each block (highest depth block which is valid and covered by this block). 
@@ -375,7 +375,7 @@ extern EbErrorType eb_vp9_sb_params_init(
         for (md_scan_block_index = EP_BLOCK_MAX_COUNT - 1; md_scan_block_index-- != 0; )
         {
             // Initialize the valid block
-            sequence_control_set_ptr->sb_params_array[sb_index].ep_scan_block_valid_block[md_scan_block_index] = (uint32_t)~0;
+            sequence_control_set_ptr->sb_params_array[sb_index].ep_scan_block_valid_block[md_scan_block_index] = (uint16_t)~0;
 
             // If this block is valid, set it as the valid block.
             if (sequence_control_set_ptr->sb_params_array[sb_index].ep_scan_block_validity[md_scan_block_index] == EB_TRUE)
@@ -388,8 +388,8 @@ extern EbErrorType eb_vp9_sb_params_init(
                 if (ep_block_stats_ptr->bsize > 3)
                 {
                     // Check all of the blocks at the lower depth and find the first one that is valid.
-                    for (int search_valid_index = md_scan_block_index; search_valid_index < (md_scan_block_index + ep_inter_depth_offset); search_valid_index++) {
-                        if (search_valid_index < EP_BLOCK_MAX_COUNT && sequence_control_set_ptr->sb_params_array[sb_index].ep_scan_block_valid_block[search_valid_index] != (uint32_t)~0) {
+                    for (uint16_t search_valid_index = md_scan_block_index; search_valid_index < (md_scan_block_index + ep_inter_depth_offset); search_valid_index++) {
+                        if (search_valid_index < EP_BLOCK_MAX_COUNT && sequence_control_set_ptr->sb_params_array[sb_index].ep_scan_block_valid_block[search_valid_index] != (uint16_t)~0) {
                             sequence_control_set_ptr->sb_params_array[sb_index].ep_scan_block_valid_block[md_scan_block_index] = sequence_control_set_ptr->sb_params_array[sb_index].ep_scan_block_valid_block[search_valid_index];
                             break;
                         }
